added wallhaven-favourites-browser-downloader.js
Instead of creating an elaborate program to download all the images, just let the browser do the heavy lifting.

This script fetches all the images in the current page and loads them into a new page of just those images.
This allows saving the page as "full webpage" so the browser does the heavy-lifting of downloading the images.

Note: Robustness wasn't the primary concern, so it doesn't check all edge-cases and could have some issues with very large sets.
It's also not set-up to be ran as a userscript or anything of that nature. It was made to be manually executed in a console or such.